2022. 03. 19
잠금 버그 수정

+++ Assets/03_Script/TileControl.cs
@@ -145,6 +145,11 @@ public class TileControl : MonoBehaviour
 
     public void MergeUnit()
     {
+        if (_target.GetUnit().IsLocked())
+            return;
+        if (_sameUnitTile.GetUnit().IsLocked())
+            return;
+
         _target.CreateUnit();
         _sameUnitTile.DeleteUnit();
         _sameUnitTile.ActiveEffect_select(false);
